Fix price not being deleted on car delete Delete maps allocation on car delete

+++ pricing-service/src/main/java/com/udacity/pricing/api/PricingController.java
@@ -41,7 +41,7 @@ public class PricingController {
     }
 
     @DeleteMapping(PRICING_DELETE_BY_VEHICLEID_URL)
-    public long deleteByVehicleId(@RequestParam Long vehicleId) {
+    public int deleteByVehicleId(@RequestParam Long vehicleId) {
         return pricingService.deleteByVehicleId(vehicleId);
     }
 }
